Autowait's implicit taskwait: do not finalize early released accesses
This fixes nasty_test3_taskwait_noflush4 and namespace_namespacenewfail17.

Autowait means that the accesses that are propagated in the namespace can be
early released to the successor task, but the accesses that need to be released
back to the offloader are delayed (which gives an opportunity to defragment
these accesses and reduce the number of messages). The delayed accesses are
delayed using an implicit taskwait.  This implicit taskwait does not involve
the early released accesses, so when it completes these early released accesses
may not yet be complete (either because subtasks haven't finished or
satisfiability hasn't been received). In this case the implicit taskwait should
not finalize all the fragments (which is what it was previously doing). It now
finalizes only the fragments corresponding to accesses that are not early
released in the namespace.

+++ src/dependencies/linear-regions-fragmented/DataAccessRegistration.cpp
@@ -3074,6 +3074,14 @@ namespace DataAccessRegistration {
 
 				DataAccessStatusEffects initialStatus(accessOrFragment);
 
+				// If the access is early released in the namespace (autowait
+				// feature), then set early release in the fragments too.
+				if (dataAccess->getEarlyReleaseInNamespace()) {
+					if (accessOrFragment != dataAccess) {
+						accessOrFragment->setEarlyReleaseInNamespace();
+					}
+				}
+
 				// the access will already be complete only if it is a task with a wait clause
 				if (accessOrFragment->complete()) {
 					return true;
@@ -3365,7 +3373,7 @@ namespace DataAccessRegistration {
 		} else {
 			/* With autowait, only create taskwait fragments for bottom map
 			 * entries that have not already been early released in the
-			 * namespace (by unregisterNamespaceTaskDataAccesses). We have to
+			 * namespace (by unregisterLocallyPropagatedTaskDataAccesses). We have to
 			 * ensure that bottom map entries not covered by any access still
 			 * get a taskwait fragment, since otherwise there would be a hang
 			 * on incorrect programs where the parent pragmas don't cover all
@@ -4629,8 +4637,10 @@ namespace DataAccessRegistration {
 					// for a task with a wait clause, then the access may be discounted
 					// assert(!dataAccess->hasBeenDiscounted());
 
-					if (dataAccess->hasSubaccesses()) {
-						dataAccess->unsetHasSubaccesses();
+					if (!dataAccess->getEarlyReleaseInNamespace()) {
+						if (dataAccess->hasSubaccesses()) {
+							dataAccess->unsetHasSubaccesses();
+						}
 					}
 
 					/* continue, to process all accesses */
@@ -4647,15 +4657,15 @@ namespace DataAccessRegistration {
 				[&](TaskDataAccesses::access_fragments_t::iterator position) -> bool {
 					DataAccess *dataAccess = &(*position);
 					assert(dataAccess != nullptr);
-
-					Instrument::removedDataAccess(dataAccess->getInstrumentationId());
-					accessStructures._accessFragments.erase(dataAccess);
-					ObjectAllocator<DataAccess>::deleteObject(dataAccess);
+					if (!dataAccess->getEarlyReleaseInNamespace()) {
+						Instrument::removedDataAccess(dataAccess->getInstrumentationId());
+						accessStructures._accessFragments.erase(dataAccess);
+						ObjectAllocator<DataAccess>::deleteObject(dataAccess);
+					}
 
 					/* continue, to process all access fragments */
 					return true;
 				});
-			accessStructures._accessFragments.clear();
 		}
 
 		unfragmentTaskwaits(accessStructures);
